requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
streamlit==1.36.0
openai==0.28.0
python-dotenv==1.0.1
requests==2.32.3
pydantic==2.10.6
pydantic-settings>=2.1.0
mlflow==2.4.0
anthropic==0.49.0
ruff==0.9.10